requirements.txt
asgiref==3.3.4
cffi==1.14.5
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.15
django-cors-headers==3.10.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.3.1
psycopg2==2.9.1
pycparser==2.20
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0

# Floran Todo.
<p align="center" width="100%">
    <img src="https://github.com/bhagwanZaki/react-dj-todo/blob/main/assests/logo.png"> 
</p>

## Intoduction 

Floran todo is todo app that will help you to manage your daily tasks and give you a graphical representation of task completed every day.
<br> 
<b>Stack used for this application</b>
<ul>
    <li>Frontend : React
    <li> Backend api : Django rest framework
    <li> Mobile App : Flutter
    <li>Database : PostgresSql 
</ul>


## Feature Available 

<ul>
    <li>  Graphical Represetation of task completed per day
    <li>User Registration
    <li> Add/Delete/Complete Todo
   
</ul>

## Pictures

<p align="center" width="100%">
    <h2> Welcome Page </h2>
    <img src="https://github.com/bhagwanZaki/react-dj-todo/blob/main/assests/main%20page.png"> 
    <h2> Mainpage </h2>
    <img src="https://github.com/bhagwanZaki/react-dj-todo/blob/main/assests/dashboard.png"> 
</p>


## Requirement 

Python v.3.8+
<br>Node.JS <br>Flutter v.2.2.1


  
